splitted VS moduels interfaces and done typings
